epwa2707it6/line02/requirements.txt
line-bot-sdk
Flask==1.0.2